Fixed salesTax.js by wrapping everthing in a <form> tag IT WORKS NOW

+++ src/Helper.js
@@ -2,7 +2,7 @@ function totalPrice(arr) {
   const num = arr.map(item => {
     return {
       name: item.name,
-      price: parseInt(item.price)
+      price: parseFloat(item.price)
     };
   });
   return num.reduce((curr, item) => curr + item.price, 0);
